test: keep durable lifecycle tests under size limits
