+++ pkg/services/factory_sessions/internal/execution/javascript_runtime_start_test.go
@@ -380,188 +380,6 @@ func TestJavaScriptRuntimeService_StartAsync_RunningCancelAndReads(t *testing.T)
 	}
 }
 
-func TestJavaScriptRuntimeService_CloseCancelsJoinsAndPersistsAsyncSession(t *testing.T) {
-	t.Parallel()
-
-	projectRoot := t.TempDir()
-	store := mustTestRuntimePersistenceStore(t, runtimepersist.DirForProjectRoot(projectRoot))
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: projectRoot,
-		Persistence: store,
-		Workflows:   scriptedBlockingRuntimeWorkflows(),
-	})
-	started, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-close-joins-001",
-		busyLoopWorkflowSource,
-		map[string]any{"subject": "shutdown"},
-		nil,
-	))
-	if err != nil {
-		t.Fatalf("StartAsync: %v", err)
-	}
-
-	if err := service.Close(); err != nil {
-		t.Fatalf("Close: %v", err)
-	}
-
-	session, err := service.GetSession(context.Background(), started.SessionID)
-	if err != nil {
-		t.Fatalf("GetSession after Close: %v", err)
-	}
-	if session.Status != LifecycleStatusCanceled {
-		t.Fatalf("session status after Close = %q, want CANCELED", session.Status)
-	}
-	if session.Failure == nil || session.Failure.Reason != "WORKFLOW_RUNTIME_CANCELED" {
-		t.Fatalf("session failure after Close = %#v, want WORKFLOW_RUNTIME_CANCELED", session.Failure)
-	}
-	snapshotPath := filepath.Join(runtimepersist.DirForProjectRoot(projectRoot), started.SessionID+".json")
-	if _, err := os.Stat(snapshotPath); err != nil {
-		t.Fatalf("terminal snapshot after Close: %v", err)
-	}
-	if err := service.Close(); err != nil {
-		t.Fatalf("repeated Close: %v", err)
-	}
-	if _, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-close-rejected-001",
-		busyLoopWorkflowSource,
-		nil,
-		nil,
-	)); !errors.Is(err, ErrDurableExecutionClosed) {
-		t.Fatalf("StartAsync after Close error = %v, want ErrDurableExecutionClosed", err)
-	}
-}
-
-func TestJavaScriptRuntimeService_CloseRejectsStartWithoutOrphaningReservation(t *testing.T) {
-	t.Parallel()
-
-	t.Run("async", func(t *testing.T) {
-		exerciseStartAdmissionAfterClose(t, func(service *JavaScriptRuntimeService) error {
-			_, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-				"req-runtime-close-admission-async-001",
-				simpleFinalWorkflowSource,
-				map[string]any{"subject": "shutdown"},
-				nil,
-			))
-			return err
-		})
-	})
-
-	t.Run("wait sync", func(t *testing.T) {
-		waitMillis := int64(100)
-		exerciseStartAdmissionAfterClose(t, func(service *JavaScriptRuntimeService) error {
-			request := inlineWorkflowStartRequest(
-				"req-runtime-close-admission-sync-001",
-				simpleFinalWorkflowSource,
-				map[string]any{"subject": "shutdown"},
-				nil,
-			)
-			request.Wait = &WaitOptions{TimeoutMillis: &waitMillis}
-			_, err := service.StartSync(context.Background(), request)
-			return err
-		})
-	})
-}
-
-func exerciseStartAdmissionAfterClose(
-	t *testing.T,
-	start func(*JavaScriptRuntimeService) error,
-) {
-	t.Helper()
-	entered := make(chan struct{})
-	release := make(chan struct{})
-	workflows := &blockingWorkflowDefinitions{
-		JavaScriptWorkflows: scriptedSuccessfulRuntimeWorkflows(map[string]any{"status": "started"}),
-		entered:             entered,
-		release:             release,
-	}
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: t.TempDir(),
-		Workflows:   workflows,
-	})
-	startDone := make(chan error, 1)
-	go func() { startDone <- start(service) }()
-
-	select {
-	case <-entered:
-	case <-time.After(5 * time.Second):
-		t.Fatal("timed out waiting for start preparation")
-	}
-	if err := service.Close(); err != nil {
-		t.Fatalf("Close: %v", err)
-	}
-	close(release)
-	if err := <-startDone; !errors.Is(err, ErrDurableExecutionClosed) {
-		t.Fatalf("start error = %v, want ErrDurableExecutionClosed", err)
-	}
-
-	service.mu.RLock()
-	sessionCount := len(service.sessions)
-	replayCount := len(service.startReplay)
-	inflightCount := len(service.startInflight)
-	service.mu.RUnlock()
-	if sessionCount != 0 || replayCount != 0 || inflightCount != 0 {
-		t.Fatalf(
-			"close-rejected start left durable state: sessions=%d replay=%d inflight=%d",
-			sessionCount, replayCount, inflightCount,
-		)
-	}
-}
-
-type blockingWorkflowDefinitions struct {
-	factory.JavaScriptWorkflows
-	entered chan struct{}
-	release chan struct{}
-}
-
-func (workflows *blockingWorkflowDefinitions) ResolveSource(
-	request factory.WorkflowSourceRequest,
-	context factory.WorkflowSourceContext,
-) factory.WorkflowSourceResolution {
-	close(workflows.entered)
-	<-workflows.release
-	return workflows.JavaScriptWorkflows.ResolveSource(request, context)
-}
-
-func TestJavaScriptRuntimeService_CloseReportsNonCooperativeRunTimeout(t *testing.T) {
-	t.Parallel()
-
-	entered := make(chan struct{})
-	release := make(chan struct{})
-	service := newConfiguredJavaScriptRuntimeService(javaScriptRuntimeServiceConfig{
-		ProjectRoot: t.TempDir(),
-		Workflows: scriptedRuntimeWorkflows(func(
-			context.Context,
-			factory.JavaScriptRuntimeRequest,
-			factory.JavaScriptRuntimeHooks,
-		) (factory.JavaScriptRuntimeOutcome, error) {
-			close(entered)
-			<-release
-			return factory.JavaScriptRuntimeOutcome{OK: true}, nil
-		}),
-	})
-	defer func() {
-		close(release)
-		service.runWaitGroup.Wait()
-	}()
-	if _, err := service.StartAsync(context.Background(), inlineWorkflowStartRequest(
-		"req-runtime-close-timeout-001",
-		busyLoopWorkflowSource,
-		nil,
-		nil,
-	)); err != nil {
-		t.Fatalf("StartAsync: %v", err)
-	}
-	select {
-	case <-entered:
-	case <-time.After(time.Second):
-		t.Fatal("timed out waiting for the non-cooperative workflow to start")
-	}
-
-	if err := service.closeWithTimeout(time.Millisecond); !errors.Is(err, ErrDurableExecutionShutdownTimeout) {
-		t.Fatalf("closeWithTimeout error = %v, want ErrDurableExecutionShutdownTimeout", err)
-	}
-}
-
 func TestJavaScriptRuntimeService_StartAsync_FailedAndTimedOut(t *testing.T) {
 	t.Parallel()
 	t.Run("failed", func(t *testing.T) {
